+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
-## 1.0.0 (Unreleased)
+## 1.0.0 (2024-04-01)
 
 Initial release.
+
+## 1.0.1 (Unreleased)
+
+Fixes for the Grafana plugin.
+Grafana plugin publishing token.